evaluation/src/ERQA-main/requirements.txt
tensorflow>=2.8.0
numpy>=1.22.0
pillow>=9.0.0
google-genai>=0.3.0
openai>=1.0.0
requests>=2.28.0
argparse>=1.4.0 
